% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_quad_count}
\alias{get_quad_count}
\title{Get the count of events based on Quad category}
\usage{
get_quad_count(conn, quad, date = NULL, source.country = NULL,
  target.country = NULL)
}
\arguments{
\item{conn}{A MySQL connection to the ICEWS database. Supply your credentials
to \code{conn <- dbConnect(MySQL(), user="username", password="password",
dbname="event_data", host="our.host.ip")}}

\item{quad}{A string indicates a quad category, i.e. one of
\code{verb.coop, verb.conf, matl.coop, matl.confl)}}

\item{date}{A string in the YYYYMMDD format. If \code{NULL}, will count events on
any date}

\item{source.country}{A string specifying the ISOA3Code of the source country.
If \code{NULL}, will count events from all countries}

\item{target.country}{A string specifying the ISOA3Code of the target country.
If \code{NULL}, will count events targeted at all countries}
}
\description{
\code{get_quad_count} takes a vector of quad categories
(including verbal/material cooperation/conflict), determines the corresponding
CAMEO codes, then get the counts for the specified source/target countries
and date by calling \code{get_CAMEO_count}
}
\examples{
\dontrun{
quads <- c("verb.coop", "verb.conf", "matl.coop", "matl.conf")
# Get quad counts between all country-dyads in a given date
result1 <- ldply(quads, .fun=getIcewsQuad, date="20021025")
# Get quad counts between one dyad for all dates
result2 <- ldply(quads, .fun=getIcewsQuad,
                source.country="AFG", target.country="ALB")
# Get quad counts for one dyad in one date
result3 <- ldply(quads, .fun=getIcewsQuad, date="20021025", source.country="AFG", target.country="ALB")
}
}

