% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{icewsr}
\alias{icewsr}
\alias{icewsr-package}
\title{Utility functions to query and manipulate data from ICEWS}
\description{
This package includes various utility functions that I have
created to query and manipulate data from the ICEWS database
}

